coupon-system/src/main/java/com/example/coupon_system/service/ApplyService.java
package com.example.coupon_system.service;

import com.example.coupon_system.domain.Coupon;
import com.example.coupon_system.repository.CouponRepository;
import org.springframework.stereotype.Service;

@Service
public class ApplyService {

    private final CouponRepository couponRepository;

    public ApplyService(CouponRepository couponRepository) {
        this.couponRepository = couponRepository;
    }

    public void apply(Long userId) {
        long count = couponRepository.count();

        if (count > 100) {
            return;
        }

        couponRepository.save(new Coupon(userId));
    }

}
