coupon-system/src/test/java/com/example/coupon_system/service/ApplyServiceTest.java
package com.example.coupon_system.service;

import static org.assertj.core.api.Assertions.assertThat;

import com.example.coupon_system.repository.CouponRepository;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

@SpringBootTest
class ApplyServiceTest {

    @Autowired
    private ApplyService applyService;

    @Autowired
    private CouponRepository couponRepository;


    @Test
    void 한번응모() {
        applyService.apply(1L);

        long count = couponRepository.count();

        assertThat(count).isEqualTo(1);
    }
}
